requirements.txt
yt-dlp
PyQt6
mutagen
musicbrainzngs
pycdlib
google-generativeai